Move integrate into isolate.js
