+++ tools/isolate/index.js
@@ -3,8 +3,7 @@ const path = require('path');
 const glob = util.promisify(require('glob'));
 const mv = require('move-file');
 
-async function run() {
-  const ignore = process.argv.slice(2);
+async function isolate(ignore = []) {
   const matches = await glob('src/site/content/**/*.md', {ignore});
   matches.forEach(async (oldPath) => {
     let newPath = oldPath.split(path.sep);
@@ -14,4 +13,25 @@ async function run() {
   });
 }
 
-run();
+async function integrate() {
+  const matches = await glob('src/site/_exile/**/*.md');
+  matches.forEach(async (oldPath) => {
+    let newPath = oldPath.split(path.sep);
+    newPath.splice(2, 1, 'content');
+    newPath = newPath.join(path.sep);
+    await mv(oldPath, newPath);
+  });
+}
+
+const command = process.argv[2];
+if (command !== 'isolate' && command !== 'integrate') {
+  throw new Error(
+    `Invalid command: ${command}. Command must be 'isolate' or 'integrate'.`,
+  );
+}
+const ignore = process.argv.slice(3);
+if (command === 'isolate') {
+  return isolate(ignore);
+} else if (command === 'integrate') {
+  return integrate();
+}
